requirements-monitor.txt
paho-mqtt==1.4.0
pkg-resources==0.0.0
pycairo==1.17.1
pydbus==0.6.0
PyGObject==3.30.1
